requirements.txt
# =============================================================================
# Super Manager - Python Dependencies
# =============================================================================

# ===== Core Framework =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# ===== AI Providers =====
openai==1.3.0
groq==0.4.0
httpx==0.25.1
# Note: Ollama uses HTTP API, no SDK needed

# ===== Database =====
# Supabase (Primary)
supabase==2.3.0

# Firebase (Legacy - optional)
google-cloud-firestore==2.13.1
google-auth==2.25.2

# ===== Task Queue & Caching =====
celery==5.3.4
redis==5.0.1

# ===== Security =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== LangChain (Optional - for advanced chains) =====
langchain==0.0.350
langchain-openai==0.0.2

# ===== Communication =====
# Email is built-in with smtplib
# Telegram uses httpx

# ===== Monitoring & Logging =====
prometheus-client==0.19.0

# ===== Development =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
